Delete bindings.rs file
This file is stale; the up to date copy is generated each time you build
in the build output directory.

Leaving this file around may lead a curious repo-goer to try and read or
understand it when in reality it's not useful.

Alternatively, we could keep an up-to-date file in here in any of a
number of ways, but it doesn't seem like the normal way things are done
in rust.